// SETPPBO/ClientApp/app/main/referensi/models/referensi-majelis.ts
export class ReferensiMajelis {
    RefMajelisId: number;
    Kode: string;
    Majelis: string;
    Harsinom: string;
    HakimKetuaId: number;
    NamaHakimKetua: string;
    NiphakimKetua: string;
    HakimAnggota1Id: number;
    NamaHakimAnggota1: string;
    NiphakimAnggota1: string;
    HakimAnggota2Id: number;
    NamaHakimAnggota2: string;
    NiphakimAnggota2: string;
    Spid: number;
    NamaSp: string;
    Nipsp: string;
    Psp1id: number;
    Psp2id: number;
    RefJenisKasusId: number;
    TotalBerkas: number;
    // RefJenisKasusID: number;
    // Uraian: string;
}